+++ FreeRTOS/Demo/R-Car_Gen5_CR52/drivers/taud/r_taud_api.c
@@ -53,7 +53,7 @@ tau_clock_t * taud_pwm_set_clk(uint8_t unit, uint32_t freq_hz)
             clk_set = &tau_clock[ck_index];
             for (uint8_t presc = 0; presc <= TAUD_PRESCALER_MAX; presc++)
             {
-                clk_set->clk_value = scpbusd8_scp_main / (1U << presc); 
+                clk_set->clk_value = scpbusd8_scp_main / (uint32_t)(1U << presc); 
                 if(clk_set->clk_value / freq_hz < TAUD_COUNTER_MAX_TICKS)
                 {
                     clk_set->clk_src = ck_index;
@@ -65,7 +65,7 @@ tau_clock_t * taud_pwm_set_clk(uint8_t unit, uint32_t freq_hz)
 
             if(clk_set->clk_src == TAUD_CLK_CK3)
             {
-                clk_set->clk_value = scpbusd8_scp_main / (1U << TAUD_PRESCALER_MAX);
+                clk_set->clk_value = scpbusd8_scp_main / (uint32_t)(1U << TAUD_PRESCALER_MAX);
                 taud_pwm_set_prescaler(unit, clk_set->clk_src, TAUD_PRESCALER_MAX);
 
                 uint32_t ck3_presc = clk_set->clk_value / TAUD_COUNTER_MAX_TICKS;
